added ohnesorgeApproach

+++ src/main/java/de/ohnes/util/Job.java
@@ -5,16 +5,17 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Getter;
 import lombok.Setter;
 
-@Getter @Setter
+@Getter
+@Setter
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Job {
-    
+
     private long id;
     private int[] processingTimes;
     private int startingTime;
     private int allotedMachines;
 
-    public Job(int id, int[] processingTimes) {
+    public Job(long id, int[] processingTimes) {
         this.id = id;
         this.processingTimes = processingTimes;
 
@@ -35,22 +36,27 @@ public class Job {
     }
 
     /**
-     * find out the cononical number of machines for a job with max. execution time @param h.
+     * find out the cononical number of machines for a job with max. execution
+     * time @param h.
      * using binary search
      * O(log m)
+     * 
      * @return -1 if the job cant be executed in time h
      */
     public int canonicalNumberMachines(double h) {
         int r = processingTimes.length - 1;
-        if(this.processingTimes[r] > h) {
+        if (this.processingTimes[r] > h) {
             return -1;
         }
         int l = 0;
         while (r >= l) {
             int mid = l + (r - l) / 2;
-            if (processingTimes[mid] == h) break;
-            if (processingTimes[mid] < h) r = mid - 1;
-            if (processingTimes[mid] > h) l = mid + 1;
+            if (processingTimes[mid] == h)
+                break;
+            if (processingTimes[mid] < h)
+                r = mid - 1;
+            if (processingTimes[mid] > h)
+                l = mid + 1;
         }
         return l + (r - l) / 2 + 1;
     }
@@ -59,5 +65,5 @@ public class Job {
         this.startingTime = 0;
         this.allotedMachines = 0;
     }
-    
+
 }
